requirements.txt
opencc-python-reimplemented
charset-normalizer
